compat: Reword around incompatible definitions of the `legacy_callback_t` & Globally Replace `mData` with `mUser`
This build failure for libaudioclient_shim in LineageOS compatibility layer (hardware/lineage/compat)
is due to inconsistent or incompatible definitions of the legacy_callback_t type and issues
with class inheritance/pointer handling between the shim and the underlying Android framework headers.

The second argument should be the user's opaque data pointer,
which we had correctly declared as mUser in the previous step.
The name mData is simply a typo/legacy name that needs replacement.

Change-Id: 3cd2ba1896e2ee101a84270a45e74fb657f6e06c
Signed-off-by: Shakib <dev@shakib.my.id>